feat: modifier recipes refactor finish
